Performance Optimization: De-duplicating Derived Function Traces
We collect the dependencies of a Derived Function as a
Vector{DependencyKey}. It's important that we maintain the _order_ of
the dependencies, because we need to check their validity in order when
there are changes, since a change in an earlier dependecy may invalidate
the _keys_ of a later dependency (consider how deleting an item from a
set affects an aggregating derived function such as `sum()`).

However, as a performance optimization, we do not need to keep
_duplicate_ keys, so we also maintain a Set{DependencyKey} for
deduplicating the keys as they're recorded. Using an unordered,
hash-based Set + a Vector together provides the best of both worlds.
The set should be unordered and hash-based because we want quick
membership verification (in O(1)-time), which `Set` is, and is discarded
at the end of recording the dependencies.

This commit changes dependency tracking to maintain a Vector AND a Set
used to deduplicate the dependencies as they're recorded.

----------------

For example, this commit reduced the memory usage of our Salsa cache
from `1.9 MiB` down to `1.2 MiB`.