Reduce Salsa allocs: Disable Salsa stacks if !debug_enabled()
Since the Salsa stack traces are implemented via a linked-list (in order
to support branching for calls on new threads), they necessarily
allocate (2 allocs) on each `@derived function` call. HOWEVER, it's
worth noting that this allocation is _extremely_ cheap. So we might
consider re-enabling this, even in debug mode, since it really is nice
to do early cycle detection and pretty stack trace printing.