Disable return_type inference in Salsa b/c it's hanging in Delve.
We added return_type inference to make derived functions type stable,
since julia of course has no way to know what types of values will be
found in the salsa caches.

However, this appears to never terminate (or at least, times out after
an hour on the build farm), so this commit disables it. We'll have to
rely on user-specified return type annotations for now, despite the
negative affect those can have on performance.