avoid potential compiler-internal error
It's generally better to avoid constructing excessively large tuple
since it may cause the internal error in the compiler
(see JuliaLang/julia/#38364). In particular, it turns out that this line
can cause the stack-overflow error in the optimizer (xref:
JuliaLang/julia#48913):
<https://github.com/RelationalAI-oss/Salsa.jl/blob/e8029cc8401734e5bd26e142b653ed368499f2c5/test/Salsa.jl#L386-L387>

This commit decreases the value of `I`, because our compiler does not
support a code containing such a big tuple.